requirements.txt
# Python Telegram Bot
python-telegram-bot>=20.0
